requirements.txt
prettytable
argparse
progressbar2